+++ bfabric_app_runner/src/bfabric_app_runner/specs/inputs/bfabric_order_fasta_spec.py
@@ -16,6 +16,7 @@ class BfabricOrderFastaSpec(BaseModel):
     id: int
     entity: Literal["workunit", "order"]
     filename: RelativeFilePath
+    required: bool = False
 
     def resolve_filename(self, client: Bfabric) -> str:
         return self.filename